fixing bookmark edition

+++ src/main/java/com/mageddo/bookmarks/entity/TagEntity.java
@@ -5,10 +5,10 @@ import org.springframework.jdbc.core.RowMapper;
 
 public class TagEntity {
 
-	private Long id;
+	private Integer id;
 	private String slug;
 	private String name;
-	private Long bookmarkId;
+	private Integer bookmarkId;
 
 	public String getName() {
 		return name;
@@ -19,11 +19,11 @@ public class TagEntity {
 		return this;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public TagEntity setId(Long id) {
+	public TagEntity setId(Integer id) {
 		this.id = id;
 		return this;
 	}
@@ -37,21 +37,21 @@ public class TagEntity {
 		return this;
 	}
 
-	public Long getBookmarkId() {
+	public Integer getBookmarkId() {
 		return bookmarkId;
 	}
 
-	public TagEntity setBookmarkId(Long bookmarkId) {
+	public TagEntity setBookmarkId(Integer bookmarkId) {
 		this.bookmarkId = bookmarkId;
 		return this;
 	}
 
 	public static RowMapper<TagEntity> mapper() {
 		return (rs, rowNum) -> new TagEntity()
-			.setId(rs.getLong("IDT_TAG"))
+			.setId(rs.getInt("IDT_TAG"))
 			.setName(rs.getString("NAM_TAG"))
 			.setSlug(rs.getString("COD_SLUG"))
-			.setBookmarkId(JdbcHelper.getLong(rs, "IDT_BOOKMARK"))
+			.setBookmarkId(JdbcHelper.getInteger(rs, "IDT_BOOKMARK"))
 		;
 	}
